+++ test/opwf.jl
@@ -8,6 +8,11 @@
         result = solve_opwf(p_file, w_file, link_file, pwm_type, nlp_solver; relax_integrality = true)
         @test result["termination_status"] == LOCALLY_SOLVED
         @test isapprox(result["objective"], 2932.00, rtol = 1.0e-2)
+
+        pwm_type = PowerWaterModel{LinDist3FlowPowerModel, CRDWaterModel}
+        result = run_opwf(p_file, w_file, link_file, pwm_type, nlp_solver; relax_integrality = true)
+        @test result["termination_status"] == LOCALLY_SOLVED
+        @test isapprox(result["objective"], 2932.00, rtol = 1.0e-2)
     end
 
     @testset "3-bus LinDist3FlowPowerModel and CRDWaterModel (Multistep)" begin
@@ -17,6 +22,11 @@
 
         pwm_type = PowerWaterModel{LinDist3FlowPowerModel, CRDWaterModel}
         result = solve_opwf(p_file, w_file, link_file, pwm_type, nlp_solver; relax_integrality = true)
+        @test result["termination_status"] == LOCALLY_SOLVED
+        @test isapprox(result["objective"], 8794.26, rtol = 1.0e-2)
+
+        pwm_type = PowerWaterModel{LinDist3FlowPowerModel, CRDWaterModel}
+        result = run_opwf(p_file, w_file, link_file, pwm_type, nlp_solver; relax_integrality = true)
         @test result["termination_status"] == LOCALLY_SOLVED
         @test isapprox(result["objective"], 8794.26, rtol = 1.0e-2)
     end