+++ CHANGELOG.md
@@ -2,6 +2,7 @@ PowerWaterModels.jl Change Log
 =======================
 
 ### Staged
+- Rename `run_` methods to `_solve` and add deprecation warnings.
 - Updates for InfrastructureModels v0.7.
 - Updates for PowerModelsDistribution v0.14.
 - Updates for WaterModels v0.9.